+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # CHANGELOG
 
-## [0.2.0-dev] - 20/06/22
+## [0.3.0-dev] - 23/06/2022
+
+### Implementations: Templates
+
+#### Added
+
+* Templates:
+    - About page template
+    - Projects page template
+    - Project detail page template
+    - Contact page template
+
+## [0.2.0-dev] - 20/06/2022
 
 ### Implementations: Templates
 
@@ -13,7 +25,7 @@
     - Base template
     - Home page template
 
-## [0.1.0-dev] - 18/06/22
+## [0.1.0-dev] - 18/06/2022
 
 ### Start Project: Basic config + Tailwind